Cinema map edits
Long hallway -> enables a "chase" scene with a janitor door or smth and can make the hallway decrease in size and spawn enemies to force player to boss.
Janitor door for hallway could also take a key found by exploration in other theaters??
Changed lobby to include a hight difference, ramp and balcony so that player gets accustomed to verticality when aiming.